Switch from backported Collections to the native (#161)
The backported Collections package is not available in a restricted env I'm trying to run in (doesn't have access to any public repo stores like maven central).
I assume this was an accident when doing an auto-complete because the only place Collections is used is for an emptyList(), which the native one does implement. This fixes that.